+++ demo_factions.html
@@ -80,11 +80,12 @@
             // jsonPath    : "http://www.edsm.net/api-v1/systems?coords=1&known=1&startdatetime="+queryDate,
             jsonPath: "json_samples/factions.json",
             withHudPanel: true,
-            hudMultipleSelect: false,
-            effectScaleSystem: [50, 10000],
-            startAnim: false,
+            withOptionsPanel: true,
+            hudMultipleSelect: true,
+            effectScaleSystem: [10, 800],
+            startAnim: true,
             showGalaxyInfos: true,
-            cameraPos: [25, 14100, -12900],
+            systemColor: '#FF0000'
         });
 
     </script>
